+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   greater than 0% failed components will be logged as warning messages, and all
   other session statuses will be logged as info messages.
 - Changed the default value of the config file option `bos.api_version` to "v2".
+- Modified logging for SAT such that multi-line log messages will now be logged
+  with consistent formatting for each line.
 
 ### Fixed
 - Fixed a bug in the `bos-operations` stage of `sat bootsys` where a Bad Request